Move message sending internals into AbstractClientConnection.
The `Connection` methods `addMessage()`, `hasMoreMessages()` and `nextMessages()` were only used internally for
`ClientConnection `implementations, through and supported by `AbstractConnection`. The methods are no longer publicly
exposed, and the implementations have been moved out of `AbstractConnection` into `AbstractClientConnection`
